+++ Carbon.Windows.Service/Functions/Get-CServiceSecurityDescriptor.ps1
@@ -3,41 +3,73 @@ function Get-CServiceSecurityDescriptor
 {
     <#
     .SYNOPSIS
-    Gets the raw security descriptor for a service.
-    
+    Gets a Windows service's security descriptor.
+
     .DESCRIPTION
-    You probably don't want to mess with the raw security descriptor.  Try `Get-CServicePermission` instead.  Much more useful.
-    
+    The `Get-CServiceSecurityDescriptor` function gets a Windows service's security descriptor. Pass the service's name
+    to the `Name` parameter (wildcards *not* accepted). The function uses the Windows API's `OpenSCManager`,
+    `OpenService`, and `QueryServiceObjectSecurity` functions to get the service's security descriptor. Returns a
+    `[Security.AccessControl.RawSecurityDescriptor]` object with `Owner`, `Group`, and `DiscretionaryAcl` set. The
+    `SystemAcl` is not set.
+
+    User must have `ReadControl` permission to a service. Even with that permission, some services still require
+    elevated access.
+
     .OUTPUTS
     System.Security.AccessControl.RawSecurityDescriptor.
-    
+
     .LINK
     Get-CServicePermission
-    
+
     .LINK
     Grant-ServicePermissions
-    
+
     .LINK
     Revoke-ServicePermissions
-    
+
     .EXAMPLE
     Get-CServiceSecurityDescriptor -Name 'Hyperdrive'
-    
+
     Gets the hyperdrive service's raw security descriptor.
     #>
     [CmdletBinding()]
+    [OutputType([Security.AccessControl.RawSecurityDescriptor])]
     param(
-        [Parameter(Mandatory=$true)]
-        [string]
-        # The name of the service whose permissions to return.
-        $Name
+        # The name of the service whose security descriptor to return. Wildcards *not* accepted.
+        [Parameter(Mandatory)]
+        [String] $Name
     )
 
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    $sdBytes = [Carbon.Service.ServiceSecurity]::GetServiceSecurityDescriptor($Name)
-    New-Object Security.AccessControl.RawSecurityDescriptor $sdBytes,0
+    $scmHandle = Invoke-AdvApiOpenSCManager -DesiredAccess Read
+    if (-not $scmHandle)
+    {
+        return
+    }
+
+    try
+    {
+        $svcHandle = Invoke-AdvApiOpenService -SCManagerHandle $scmHandle -ServiceName $Name -DesiredAccess ReadControl
+        if (-not $svcHandle)
+        {
+            return
+        }
+
+        try
+        {
+            Invoke-AdvApiQueryServiceObjectSecurity -ServiceHandle $svcHandle `
+                                                    -SecurityInformation 'Owner, Group, DiscretionaryAcl'
+        }
+        finally
+        {
+            $svcHandle | Invoke-AdvApiCloseServiceHandle
+        }
+    }
+    finally
+    {
+        $scmHandle | Invoke-AdvApiCloseServiceHandle
+    }
 }